+++ src/datasets/domain/models/PreviewUrl.ts
@@ -1,0 +1,5 @@
+export interface PreviewUrl {
+  token: string
+  link: string
+  isAnonymizedAccess: boolean
+}