+++ src/datasets/domain/useCases/GetDataset.ts
@@ -17,19 +17,22 @@ export class GetDataset implements UseCase<Dataset> {
    * @param {string | DatasetNotNumberedVersion} [datasetVersionId=DatasetNotNumberedVersion.LATEST] - The dataset version identifier, which can be a version-specific numeric string (for example, 1.0) or a DatasetNotNumberedVersion enum value. If this parameter is not set, the default value is: DatasetNotNumberedVersion.LATEST
    * @param {boolean} [includeDeaccessioned=false] - Indicates whether to consider deaccessioned versions in the dataset search or not. The default value is false
    * @param {boolean} [keepRawFields=false] - Indicates whether or not the use case should keep the metadata fields as they are and avoid the transformation to markdown. The default value is false.
+   * @param {string} [previewUrlToken] - The token identifying a Preview URL, allowing a reviewer without credentials to access an unpublished dataset (optional).
    * @returns {Promise<Dataset>}
    */
   async execute(
     datasetId: number | string,
     datasetVersionId: string | DatasetNotNumberedVersion = DatasetNotNumberedVersion.LATEST,
     includeDeaccessioned = false,
-    keepRawFields = false
+    keepRawFields = false,
+    previewUrlToken?: string
   ): Promise<Dataset> {
     return await this.datasetsRepository.getDataset(
       datasetId,
       datasetVersionId,
       includeDeaccessioned,
-      keepRawFields
+      keepRawFields,
+      previewUrlToken
     )
   }
 }