+++ src/files/domain/useCases/GetFileAndDataset.ts
@@ -12,18 +12,21 @@ export class GetFileAndDataset implements UseCase<[FileModel, Dataset]> {
    * @param {number | string} [fileId] - The File identifier, which can be a string (for persistent identifiers), or a number (for numeric identifiers).
    * @param {string | DatasetNotNumberedVersion} [datasetVersionId=DatasetNotNumberedVersion.LATEST] - The dataset version identifier, which can be a version-specific numeric string (for example, 1.0) or a DatasetNotNumberedVersion enum value. If this parameter is not set, the default value is: DatasetNotNumberedVersion.LATEST
    * @param {boolean} [includeDeaccessioned=false] - If true, the file will be returned even if it has been deaccessioned.
+   * @param {string} [previewUrlToken] - The token identifying a Preview URL, allowing a reviewer without credentials to access an unpublished dataset (optional).
    * @returns {Promise<[FileModel, Dataset]>}
    */
   async execute(
     fileId: number | string,
     datasetVersionId: string | DatasetNotNumberedVersion = DatasetNotNumberedVersion.LATEST,
-    includeDeaccessioned = false
+    includeDeaccessioned = false,
+    previewUrlToken?: string
   ): Promise<[FileModel, Dataset]> {
     return (await this.filesRepository.getFile(
       fileId,
       datasetVersionId,
       true,
-      includeDeaccessioned
+      includeDeaccessioned,
+      previewUrlToken
     )) as [FileModel, Dataset]
   }
 }